+++ Jvedio-WPF/Jvedio.Worker/Program.cs
@@ -24,6 +24,7 @@ builder.Services.AddSingleton<LibraryTaskOrchestratorService>();
 builder.Services.AddSingleton<LibraryScanService>();
 builder.Services.AddSingleton<LibraryScrapeService>();
 builder.Services.AddSingleton<SettingsService>();
+builder.Services.AddSingleton<ActorService>();
 builder.Services.AddSingleton<VideoService>();
 builder.Services.AddSingleton<AppBootstrapService>();
 builder.Services.AddSingleton<TaskSummarySnapshotService>();